feat: move fixtures/db to github artifact (#12)
## 💸 TL;DR
So this is a messy refactor to get away from us generating fixtures and
a DB and uploading that into code (which is bad because these files are
large >10Mb and will change with some relative frequency). So to get
around that, we're going to publish these as releases and that way folks
can just download them when using this.

## 📜 Details
* Adds `.gitignore` for `./fixtures` and `./db` paths since we're
minting these as part of our artifacts now
* Update `generateChunks.ts` and `generateSqlDb.ts` to work with
ephemeral generated fixtures
* Add archive processing (create and extract) and some logging
associated with catching what's happening
* Add SIGINT signal processing since we're not appropriately cleaning up
our temp files here
* Update Github Action to publish these artifacts

## 🧪 Testing Steps / Validation
Tested this by running `node index.js` and watching the initialization
for our component of the server.

```
$ node index.js 
{"method":"notifications/message","params":{"level":"info","data":{"time":"2025-07-02T14:09:55.472Z","message":["Using release: 2025-07-02"]}},"jsonrpc":"2.0"}
{"method":"notifications/message","params":{"level":"info","data":{"time":"2025-07-02T14:09:56.715Z","message":["Database downloaded from release 2025-07-02 to: /tmp/devvit-docs.db"]}},"jsonrpc":"2.0"}
{"method":"notifications/message","params":{"level":"info","data":{"time":"2025-07-02T14:09:58.128Z","message":["Successfully extracted archive to: /tmp/fixtures"]}},"jsonrpc":"2.0"}
{"method":"notifications/message","params":{"level":"info","data":{"time":"2025-07-02T14:09:58.130Z","message":["Fixtures downloaded and extracted from release 2025-07-02 to: /tmp/fixtures"]}},"jsonrpc":"2.0"}
```

And then still able to query MCP for our tool:


![image](https://github.com/user-attachments/assets/6c8d0c77-7c31-4658-88d3-877757fff689)